Rejigger what counts as 'filled in'
Previously I was just assuming that anything in region of interest (a
rectangle enclosing the bubble circle) was valid to count. Based on the
included example this produced inconsistent results.

This commit improves on that simple assumption in 2 parts:

1 -> use what I call an 'innerRect' so a rect that's a bit smaller than
the whole bubble to neglect the black border

2 -> mask what counts with a circle within the bubble

part 2 This is ideal as isn't inflated based on the non-bubble area OR the
border so the count more accurately represents how humans think the
scanner should work.

It was a bit tricky to figure out (!!!) but the technique uses
open cv's built-in 'bitwise_and' to only count. I commented a
stackoverflow explanation that I found particularly helpful:

https://stackoverflow.com/questions/44333605/what-does-bitwise-and-operator-exactly-do-in-opencv

There is still left todo including having the circle size be consistent
across each of the bubbles in a question so that we're comparing apples
to apples.